requirements.txt
anyio==4.1.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
cffi==1.16.0
cryptography==41.0.7
decorator==5.1.1
exceptiongroup==1.2.0
future==0.18.3
greenlet==3.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
psycopg2==2.9.9
pycparser==2.21
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tornado==6.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
